fix: moved alert defaults and added default queries for logs and traces
